requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
requests==2.32.3
beautifulsoup4==4.12.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
faiss-cpu==1.8.0.post1
google-generativeai==0.7.2
openai==1.51.0
python-dotenv==1.0.1
httpx==0.27.2
jinja2==3.1.4
playwright==1.45.0
reportlab==4.2.5
markdown==3.6
